Install Bundler globally
So that it is available in all gemsets. Otherwise, if you clone
a repo that uses a `.ruby-gemset`, bundler won't be automatically
available and `bundle install` will fail. Beginners might not
know they can `gem install bundler` to fix the issue.
Since Bundler is widely used, it makes sense to install it in the
global gemset.

RVM stopped shipping with bundler, but we can make it install
bundler in the global gemset by passing the `--with-gems` flag.