Replace Qt with PhantomJS
Why:
The original reason for including Qt was to provide support for Ruby apps that are using capybara-webkit. However, capybara-webkit will soon require Qt5 to be installed, but Qt5 requires the full Xcode package.

I don't think it's possible to automate the installation of Xcode, and even if it were possible, I wouldn't want to force everyone to download a multi-gigabyte file. Instead, it makes more sense to recommend PhantomJS and Poltergeist (for Ruby apps), which works just as well as capybara-webkit.